+++ src/shapes/EArc.js
@@ -1,5 +1,5 @@
 /**
- * @classdesc 可以进行圆弧形填充或线框绘制的显示对象
+ * @classdesc 可以进行椭圆弧形填充或线框绘制的显示对象
  * @class 
  * @extends soya2d.DisplayObjectContainer
  * @param {Object} data 所有父类参数,以及新增参数
@@ -10,24 +10,24 @@
  * @param {String} data.endAngle 弧形的结束角度
  * @author {@link http://weibo.com/soya2d MrSoya}
  */
-soya2d.Arc = function(data){
+soya2d.EArc = function(data){
 	data = data||{};
 	soya2d.DisplayObjectContainer.call(this,data);
 	soya2d.ext(this,data);
 
     this.fillStyle = data.fillStyle || 'transparent';
 };
-soya2d.inherits(soya2d.Arc,soya2d.DisplayObjectContainer);
-soya2d.ext(soya2d.Arc.prototype,{
+soya2d.inherits(soya2d.EArc,soya2d.DisplayObjectContainer);
+soya2d.ext(soya2d.EArc.prototype,{
     onRender:function(g){
         g.beginPath();
 
         g.fillStyle(this.fillStyle);
-        var sr = (this.startAngle||0)*soya2d.Math.ONERAD,
-            er = (this.endAngle||0)*soya2d.Math.ONERAD;
-        g.arc(this.w/2,this.h/2,this.w/2,sr,er);
+        var sa = this.startAngle || 0,
+            ea = this.endAngle || 0;
+        g.eArc(this.w/2,this.h/2,this.w/2,this.h/2,sa,ea);
         
-        if(er-sr != 0 && Math.abs(this.startAngle||0 - this.endAngle||0) != 360){
+        if(ea-sa != 0 && Math.abs(sa - ea) != 360){
             g.lineTo(this.w/2,this.h/2);
         }
         g.fill();